Return a 500 error when `item.remove` fails (#4832)
* Return a 500 error when `item.remove` fails

Currently, errors from `item.remove` are not handled at all in the callback. This gives an empty 200 back to the frontend, which will do nothing with this response (no visual feedback on the error).
We should return an actual error. All suggestions on how to return this error Keystone style are welcome.

* directly send the `err` to `apiError`